@@ -1,1 +1,1 @@
-__version__ = "0.38"
+__version__ = "0.39.2"
